FIX: Prevent error message from showing in wrong context_id
In the `qChanged` observer we compare `q` from params with `q` stored in
the model, but if one is null and the other undefined, we get a false
positive. This would display the "your search term is too short" error
message to users when returning to the search screen without a `q` param.